[intel-npu] internal properties tests: separating implementation from instantiation
